# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Manga Store API",
        "description": "API sederhana untuk ecommerce toko buku manga. Mock data disimpan dalam file JSON (non-database).",
        "contact": {
            "name": "Muhammad Alvin Satria",
            "email": "[email protected]"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local development server"
        }
    ],
    "paths": {
        "/api/manga": {
            "get": {
                "tags": [
                    "Manga"
                ],
                "summary": "Menampilkan semua manga",
                "operationId": "23230af42e19e5a09fd6e4138043de38",
                "responses": {
                    "200": {
                        "description": "Daftar manga berhasil diambil",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Daftar manga berhasil diambil"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Manga"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Manga"
                ],
                "summary": "Menambahkan manga baru",
                "operationId": "90f5cabf9df769cee599d16c0cd98828",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/MangaInput"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Manga berhasil ditambahkan",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Manga berhasil ditambahkan"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Manga"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validasi gagal",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Validasi gagal"
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/manga/{id}": {
            "get": {
                "tags": [
                    "Manga"
                ],
                "summary": "Menampilkan manga berdasarkan ID",
                "operationId": "6d7893a01f25c88de849984a0535b903",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID manga",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detail manga ditemukan",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Detail manga ditemukan"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Manga"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Manga tidak ditemukan",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Manga"
                ],
                "summary": "Mengedit seluruh data manga (replace)",
                "operationId": "86db7fc5f0f8c1cd585df407aa5b59e9",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID manga",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/MangaInput"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Manga berhasil diperbarui",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Manga berhasil diperbarui"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Manga"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Manga tidak ditemukan",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validasi gagal"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Manga"
                ],
                "summary": "Menghapus manga berdasarkan ID",
                "operationId": "6f73b236f97f79c6737afc16658fd6f0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID manga",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Manga berhasil dihapus",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Manga berhasil dihapus"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Manga tidak ditemukan",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            },
            "patch": {
                "tags": [
                    "Manga"
                ],
                "summary": "Mengedit sebagian data manga (partial update)",
                "operationId": "b1216be3b396c6e96cc407c2b371f908",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID manga",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "One Piece Vol. 1 (Revisi)"
                                    },
                                    "price": {
                                        "type": "integer",
                                        "example": 40000
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Manga berhasil diperbarui sebagian",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Manga berhasil diperbarui sebagian"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Manga"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Manga tidak ditemukan",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validasi gagal"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Manga": {
                "required": [
                    "id",
                    "title",
                    "author",
                    "price",
                    "genre",
                    "stock"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "title": {
                        "type": "string",
                        "example": "One Piece Vol. 1"
                    },
                    "author": {
                        "type": "string",
                        "example": "Eiichiro Oda"
                    },
                    "price": {
                        "type": "integer",
                        "example": 35000
                    },
                    "genre": {
                        "type": "string",
                        "example": "Shounen"
                    },
                    "stock": {
                        "type": "integer",
                        "example": 15
                    }
                },
                "type": "object"
            },
            "MangaInput": {
                "required": [
                    "title",
                    "author",
                    "price",
                    "genre",
                    "stock"
                ],
                "properties": {
                    "title": {
                        "type": "string",
                        "example": "Bleach Vol. 1"
                    },
                    "author": {
                        "type": "string",
                        "example": "Tite Kubo"
                    },
                    "price": {
                        "type": "integer",
                        "example": 37000
                    },
                    "genre": {
                        "type": "string",
                        "example": "Shounen"
                    },
                    "stock": {
                        "type": "integer",
                        "example": 10
                    }
                },
                "type": "object"
            },
            "ErrorResponse": {
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": false
                    },
                    "message": {
                        "type": "string",
                        "example": "Item dengan ID 99 tidak Ditemukan"
                    }
                },
                "type": "object"
            }
        }
    },
    "tags": [
        {
            "name": "Manga",
            "description": "Manga"
        }
    ]
}